+++ packages/site/src/ui/transaction/components/display/TransactionIcon/TransactionCardIcon.types.ts
@@ -1,0 +1,9 @@
+import { Transaction } from 'xrpl';
+
+export interface TransactionCardIconProps {
+  className?: string;
+  style?: React.CSSProperties;
+  loading?: boolean;
+  txType: Transaction['TransactionType'];
+  isReceiver: boolean;
+}
